refactor : 뷰와 관련되지 않은 로직 도메인으로 이동
