+++ src/main/java/chess/App.java
@@ -1,12 +1,10 @@
 package chess;
 
 import chess.controller.ChessGameController;
-import chess.domain.ChessGame;
 
 public class App {
     public static void main(String[] args) {
-        ChessGame chessGame = new ChessGame();
-        ChessGameController chessGameController = new ChessGameController(chessGame);
+        ChessGameController chessGameController = new ChessGameController();
         chessGameController.start();
     }
 }
