+++ src/main/java/chess/controller/ChessGameController.java
@@ -2,9 +2,6 @@ package chess.controller;
 
 import chess.domain.ChessGame;
 import chess.domain.Team;
-import chess.domain.position.Col;
-import chess.domain.position.Position;
-import chess.domain.position.Row;
 import chess.view.Command;
 import chess.view.InputView;
 import chess.view.OutputView;
@@ -12,13 +9,6 @@ import chess.view.OutputView;
 import java.util.Objects;
 
 public final class ChessGameController {
-    private final ChessGame chessGame;
-    private Team currentTurn;
-
-    public ChessGameController(final ChessGame chessGame) {
-        this.chessGame = chessGame;
-        this.currentTurn = Team.WHITE;
-    }
 
     public final void start() {
         OutputView.printStartMessage();
@@ -26,58 +16,48 @@ public final class ChessGameController {
     }
 
     private void run() {
+        ChessGame chessGame = new ChessGame();
         while (chessGame.isPlaying()) {
-            turnExecute();
+            turnExecute(chessGame);
         }
         if (Objects.nonNull(chessGame.winner())) {
             OutputView.printWinner(chessGame.winner(), chessGame.getScoreByTeam(Team.BLACK), chessGame.getScoreByTeam(Team.WHITE));
         }
     }
 
-    private void turnExecute() {
+    private void turnExecute(final ChessGame chessGame) {
         try {
             Command command = Command.valueOf(InputView.getCommand());
             command.execute(chessGame);
-            interactiveCommand(command);
-            printCurrentBoard(command);
+            interactiveCommand(command, chessGame);
+            printCurrentBoard(command, chessGame);
         } catch (IllegalArgumentException e) {
             OutputView.printError(e.getMessage());
         }
     }
 
-    private void interactiveCommand(final Command command) {
+    private void interactiveCommand(final Command command, final ChessGame chessGame) {
         if (command.equals(Command.MOVE)) {
-            move();
-            currentTurn = Team.getAnotherTeam(currentTurn);
+            move(chessGame);
         }
         if (command.equals(Command.STATUS)) {
             OutputView.printEachTeamScore(chessGame.getScoreByTeam(Team.BLACK), chessGame.getScoreByTeam(Team.WHITE));
         }
     }
 
-    private void printCurrentBoard(final Command command) {
+    private void move(final ChessGame chessGame) {
+        String startPoint = InputView.getPoint();
+        String endPoint = InputView.getPoint();
+        chessGame.move(startPoint, endPoint);
+    }
+
+    private void printCurrentBoard(final Command command, final ChessGame chessGame) {
         if (command.isPrint()) {
-            printBoard();
+            printBoard(chessGame);
         }
     }
 
-    private void move() {
-        String startPoint = InputView.getPoint();
-        String endPoint = InputView.getPoint();
-
-        Position startPosition = position(startPoint);
-        Position endPosition = position(endPoint);
-        chessGame.move(startPosition, endPosition, currentTurn);
-    }
-
-    private Position position(final String point) {
-        return new Position(
-                Row.getLocation(String.valueOf(point.charAt(1))),
-                Col.getLocation(String.valueOf(point.charAt(0)))
-        );
-    }
-
-    private void printBoard() {
+    private void printBoard(final ChessGame chessGame) {
         OutputView.printBoard(chessGame.getBoard());
     }
 }
