+++ src/main/java/chess/domain/ChessGame.java
@@ -2,16 +2,20 @@ package chess.domain;
 
 import chess.domain.board.Board;
 import chess.domain.board.BoardFactory;
+import chess.domain.position.Col;
 import chess.domain.position.Position;
+import chess.domain.position.Row;
 
 public final class ChessGame {
     private Board board;
+    private Team currentTurn;
     private boolean isPlaying = true;
     private Team winner;
 
     public final void initSetting() {
         BoardFactory boardFactory = new BoardFactory();
-        board = boardFactory.getBoard();
+        this.board = boardFactory.getBoard();
+        this.currentTurn = Team.WHITE;
     }
 
     public final void end() {
@@ -26,16 +30,24 @@ public final class ChessGame {
         return board;
     }
 
-    public final void move(final Position startPoint, final Position endPoint, final Team team) {
-        board.move(startPoint, endPoint, team);
-        if (board.isEnemyKingDie(team)) {
-            winner = team;
+    public final void move(final String startPoint, final String endPoint) {
+        board.move(position(startPoint), position(endPoint), currentTurn);
+        if (board.isEnemyKingDie(currentTurn)) {
+            winner = currentTurn;
             end();
         }
+        currentTurn = Team.getAnotherTeam(currentTurn);
     }
 
     public final double getScoreByTeam(final Team team) {
         return board.scoreByTeam(team);
+    }
+
+    private Position position(final String point) {
+        return new Position(
+                Row.getLocation(String.valueOf(point.charAt(1))),
+                Col.getLocation(String.valueOf(point.charAt(0)))
+        );
     }
 
     public final Team winner() {
